+++ miataru/miataru/views/iPhone/iPhone_DevicesView.swift
@@ -461,6 +461,7 @@ struct iPhone_DevicesView: View {
             }
         } else {
             store.removeDevice(byID: deviceID)
+            PersistentDataCleanup.run()
         }
     }